+++ src/main/java/model/HttpRequestMessage.java
@@ -1,7 +1,7 @@
 package model;
 
+import service.RequestService;
 import utils.HttpParser;
-import utils.IOUtils;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -10,13 +10,15 @@ import java.util.List;
 
 public class HttpRequestMessage {
 
-    private static final String CONTENT_LENGTH_KEY = "Content-Length";
-
-    private final RequestLine requestLine;
+    private RequestLine requestLine;
 
     private HttpHeaders requestHeaders;
 
-    private String body;
+    private HttpBody body;
+
+    public HttpRequestMessage(BufferedReader bufferedReader) throws IOException {
+        this(RequestService.getHttpMessageData(bufferedReader), bufferedReader);
+    }
 
     public HttpRequestMessage(List<String> httpMessageData) throws IOException {
         this(httpMessageData, null);
@@ -37,31 +39,7 @@ public class HttpRequestMessage {
         }
 
         this.requestHeaders = new HttpHeaders(httpMessageData);
-        this.body = this.parseBody(bufferedReader, requestHeaders);
-    }
-
-    private String parseBody(BufferedReader bufferedReader, HttpHeaders requestHeaders) throws IOException {
-        if (bufferedReader == null) {
-            return null;
-        }
-
-        String contentLengthValue = requestHeaders
-                .getHeaders()
-                .get(CONTENT_LENGTH_KEY);
-        int contentLength = this.getContentLength(contentLengthValue);
-        if (contentLength == 0) {
-            return null;
-        }
-
-        return IOUtils.readData(bufferedReader, contentLength);
-    }
-
-    private int getContentLength(String contentLengthValue) {
-        if (contentLengthValue != null) {
-            return Integer.parseInt(contentLengthValue);
-        }
-
-        return 0;
+        this.body = new HttpBody(bufferedReader, requestHeaders);
     }
 
     public RequestLine getRequestLine() {
@@ -73,7 +51,7 @@ public class HttpRequestMessage {
     }
 
     public String getBody() {
-        return body;
+        return this.body.getContents();
     }
 
     public String toStringHttpMessage() {
@@ -82,7 +60,7 @@ public class HttpRequestMessage {
         value.append(this.requestLine.getInfo());
         value.append(this.requestHeaders.getInfo());
         value.append("\n");
-        value.append(body);
+        value.append(this.getBody());
         value.append("\n");
         value.append("]");
